Fix all enabled GCC, Clang, and cppcheck warnings
This corrects the warnings generated by the compilers (GCC, Clang)
as well as the static analysis tool cppcheck. It also re-enables some
warnings that were previously surpressed in CMakeLists although most
of these did not generate any new messages.

+++ src/Client.cpp
@@ -11,8 +11,8 @@
 Client::Client()
 {
   auto addr = Config::getMirror()[0];
-  socks_ = SocksClient::getCircuitTo(addr["ip"].asString(),
-                                     addr["port"].asInt(), 9150);
+  socks_ = SocksClient::getCircuitTo(
+      addr["ip"].asString(), static_cast<short>(addr["port"].asInt()), 9150);
   if (!socks_)
     Log::get().error("Unable to connect!");
 }
